% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{predict_flight_delay}
\alias{predict_flight_delay}
\title{Predict flight delays model}
\usage{
predict_flight_delay(flights_test, model)
}
\arguments{
\item{flights_test}{Data frame with test data}
}
\value{
Prediction
}
\description{
Predict flight delays model
}
\examples{
library(nycflights13)
data(flights)
set.seed(321)
ids <- sample(1:nrow(flights), size = 10000)
flights_train <- flights[ids,]
model <- train_flight_delay(flights_train)
ids <- sample(1:nrow(flights), size = 100)
flights_test <- flights[ids,]
results <- predict_flight_delay(flights_test, model)
}
