% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_features}
\alias{prep_features}
\title{Prepare Features}
\usage{
prep_features(flights_data)
}
\arguments{
\item{flights_data}{flights data frame}
}
\value{
data frame with features added
}
\description{
Prepare Features
}
