% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_target}
\alias{prep_target}
\title{Prepare Target}
\usage{
prep_target(flights_data)
}
\arguments{
\item{flights_data}{flights data frame}
}
\value{
data frame with target added
}
\description{
Prepare Target
}
