% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{train_flight_delay}
\alias{train_flight_delay}
\title{Train the flights model}
\usage{
train_flight_delay(flights_train)
}
\arguments{
\item{flights_train}{Flights training data frame}
}
\value{
model
}
\description{
Train the flights model
}
\examples{
set.seed(123)
ids <- sample(1:nrow(flights), size = 10000)
flights_train <- flights[ids,]
model <- train_flight_delay(flights_train)
}
